Небольшой допил. Не помню что сделал. :(

+++ src/Button.h
@@ -16,8 +16,8 @@
 
 class Button final {
 public:
-	Button(GPIO_TypeDef* gpio, uint8_t pinNum);
-	Button(GPIO_TypeDef* gpio, uint8_t pinNum, GPIOPuPd_TypeDef pull);
+	Button(GPIO_TypeDef* gpio, uint16_t pinNum);
+	Button(GPIO_TypeDef* gpio, uint16_t pinNum, GPIOPuPd_TypeDef pull);
 	bool get();
 private:
 	GPIO_TypeDef* port;
